fix some compile errors in network code ...

+++ src/jvmS16ServerCommunicationKtorMain/kotlin/lupos/s16network/ServerCommunicationTransferParams.kt
@@ -1,15 +1,17 @@
 package lupos.s16network
 
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.runBlocking
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.delay
 import io.ktor.network.selector.ActorSelectorManager
-import java.net.InetSocketAddress
 import io.ktor.network.sockets.aSocket
 import io.ktor.network.sockets.openReadChannel
 import io.ktor.network.sockets.openWriteChannel
+import io.ktor.utils.io.core.BytePacketBuilder
+import io.ktor.utils.io.core.ByteReadPacket
+import java.net.InetSocketAddress
 import kotlin.jvm.JvmField
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.runBlocking
 import lupos.s00misc.Coverage
 import lupos.s00misc.EGraphOperationType
 import lupos.s00misc.EIndexPattern
@@ -23,6 +25,7 @@ import lupos.s00misc.XMLElement
 import lupos.s03resultRepresentation.MyListValue
 import lupos.s03resultRepresentation.ResultSetDictionary
 import lupos.s03resultRepresentation.Value
+import lupos.s03resultRepresentation.ValueDefinition
 import lupos.s04arithmetikOperators.AOPBase
 import lupos.s04arithmetikOperators.noinput.AOPConstant
 import lupos.s04arithmetikOperators.noinput.AOPVariable
@@ -37,13 +40,13 @@ import lupos.s05tripleStore.TripleStoreLocalBase
 import lupos.s09physicalOperators.POPBase
 import lupos.s15tripleStoreDistributed.*
 
-object ServerCommunicationTransferParams{
-   fun receiveParams(packet: ByteReadPacket, query: Query): Array<AOPBase> {
+object ServerCommunicationTransferParams {
+    fun receiveParams(packet: ByteReadPacket, query: Query): Array<AOPBase> {
 /*always assume SPO*/
         var paramsF = Array<Boolean>(3) { true }
         var paramsS = Array<String>(3) { "" }
         for (i in 0 until 3) {
-            paramsF[i] = packet.readByte() != 0
+            paramsF[i] = packet.readByte() != 0.toByte()
             paramsS[i] = packet.readText()
         }
         val params = Array<AOPBase>(3) {
